+++ examples/activities/src/main.rs
@@ -71,8 +71,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             let resolved_emails = cx.state.resolved_emails.clone();
 
             async move {
-                let email: String = cx.call::<LookupEmail, _>("lookup", &user_id).await?;
-                cx.emit::<TrackSignupMetric, _>("track-signup", &email).await?;
+                let email: String = cx.call::<LookupEmail, _>(&user_id).await?;
+                cx.emit::<TrackSignupMetric, _>(&email).await?;
                 resolved_emails.lock().await.push(email);
                 To::done()
             }